requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
pydantic[email]
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
asyncpg>=0.29.0
alembic>=1.13.0
python-dotenv>=1.0.0
pydantic-settings>=2.1.0
# Dependencias para el agente financiero
google-genai>=0.3.0
google-generativeai>=0.8.0
websockets>=12.0
tenacity>=8.0.0
typing-inspection>=0.4.0 

# Dependencias requeridas por porfolio analizer v2/analizer_script.py
yfinance>=0.2.40
pandas>=2.2.2
numpy>=1.26.4
matplotlib>=3.8.0
plotly>=5.22.0
scipy>=1.11.0
# Exportación de imágenes con plotly (opcional pero recomendado)
kaleido>=0.2.1
# Importaciones presentes en el script
ipython>=8.20.0
streamlit>=1.33.0
tabulate>=0.9.0

# Analítica de Portfolio (para reusar Portfolio_analizer/src desde este backend)
quantstats>=0.0.62
PyPortfolioOpt>=1.5.5